pyproject.toml
[tool.poetry]
name = "fastapi-task-manager"
version = "0.1.0"
description = ""
authors = ["shatlykgurdov <gurdogly@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.0"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
alembic = "^1.16.3"
pydantic = {extras = ["email"], version = "^2.11.7"}
python-dotenv = "^1.1.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
pydantic-settings = "^2.10.1"
python-multipart = "^0.0.20"
httpx = "^0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
httpx = {version = ">=0.24.1", extras = ["http2"]}
asgi-lifespan = "^2.1.0"
anyio = "^4.4.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

